# src/swagger/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Fortress API",
		"version": "1.0.0",
		"description": "Fortress is a social media platform that allows men to connect, share experiences, and support each other in a safe and empowering environment. The platform is designed to foster positive masculinity, mental health awareness, and community engagement through various features such as posts, comments, and user profiles. This is achieved through anonimity, robust moderation, and a focus on constructive interactions."
	},
	"paths": {
...
		"/comment/post/{postId}": {
			"post": {
				"tags": ["Comment"],
				"summary": "Create comment on post",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Comment created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"get": {
				"tags": ["Comment"],
				"summary": "Get comments for a post",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "parentCommentId",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Comments list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentsListResponse"
								}
							}
						}
					}
				}
			}
		},
		"/comment/{id}": {
			"get": {
				"tags": ["Comment"],
				"summary": "Get comment by id",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Comment",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Comment"],
				"summary": "Update comment",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCommentDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"delete": {
				"tags": ["Comment"],
				"summary": "Delete comment",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericMessageResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/comment/{id}/replies": {
			"get": {
				"tags": ["Comment"],
				"summary": "Get replies to a comment",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					}
				],
				"responses": {
					"200": {
						"description": "Replies list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentsListResponse"
								}
							}
						}
					}
				}
			}
		},
		"/follow/{followingId}": {
			"post": {
				"tags": ["Follow"],
				"summary": "Follow a user",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "followingId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Follow created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FollowResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request or already following",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"delete": {
				"tags": ["Follow"],
				"summary": "Unfollow a user",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "followingId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Unfollowed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericMessageResponse"
								}
							}
						}
					},
					"404": {
						"description": "Follow not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/follow/followers/{userId}": {
			"get": {
				"tags": ["Follow"],
				"summary": "Get followers for a user",
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "status",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["pending", "accepted"]
						}
					}
				],
				"responses": {
					"200": {
						"description": "Followers list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FollowersListResponse"
								}
							}
						}
					}
				}
			}
		},
		"/follow/following/{userId}": {
			"get": {
				"tags": ["Follow"],
				"summary": "Get following for a user",
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "status",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["pending", "accepted"]
						}
					}
				],
				"responses": {
					"200": {
						"description": "Following list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FollowingListResponse"
								}
							}
						}
					}
				}
			}
		},
		"/follow/requests/pending": {
			"get": {
				"tags": ["Follow"],
				"summary": "Get pending follow requests for current user",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					}
				],
				"responses": {
					"200": {
						"description": "Pending requests",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FollowersListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/follow/requests/{followerId}/accept": {
			"post": {
				"tags": ["Follow"],
				"summary": "Accept follow request",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "followerId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Accepted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FollowResponse"
								}
							}
						}
					},
					"404": {
						"description": "Request not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/follow/requests/{followerId}/reject": {
			"delete": {
				"tags": ["Follow"],
				"summary": "Reject follow request",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "followerId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Rejected",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericMessageResponse"
								}
							}
						}
					},
					"404": {
						"description": "Request not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
...
		"/post/{id}": {
			"get": {
				"tags": ["Post"],
				"summary": "Get post by id",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Post",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Post"],
				"summary": "Update post",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePostDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Post updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Post"],
				"summary": "Delete post",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericMessageResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"okay": {
						"type": "boolean"
					}
				}
			},
			"ValidationErrorResponse": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"error": {
						"type": "string"
					},
					"okay": {
						"type": "boolean"
					}
				}
			},
			"CreateCommentDto": {
				"type": "object",
				"required": ["text"],
				"properties": {
					"text": {
						"type": "string"
					},
					"parentCommentId": {
						"type": "string"
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
			"UpdateCommentDto": {
				"type": "object",
				"required": ["text"],
				"properties": {
					"text": {
						"type": "string"
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
			"Comment": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"postId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"userAvatar": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"parentCommentId": {
						"type": "string"
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"CommentsListResponse": {
				"type": "object",
				"properties": {
					"comments": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Comment"
						}
					},
					"pagination": {
						"type": "object",
						"properties": {
							"page": {
								"type": "integer"
							},
							"limit": {
								"type": "integer"
							},
							"total": {
								"type": "integer"
							},
							"totalPages": {
								"type": "integer"
							}
						}
					}
				}
			},
			"CommentResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/Comment"
					}
				}
			},
			"GenericMessageResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					}
				}
			},
			"Follow": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"followerId": {
						"type": "string"
					},
					"followingId": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": ["pending", "accepted"]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"FollowResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/Follow"
					}
				}
			},
			"FollowersListResponse": {
				"type": "object",
				"properties": {
					"followers": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Follow"
						}
					},
					"pagination": {
						"type": "object",
						"properties": {
							"page": {
								"type": "integer"
							},
							"limit": {
								"type": "integer"
							},
							"total": {
								"type": "integer"
							},
							"totalPages": {
								"type": "integer"
							}
						}
					}
				}
			},
			"FollowingListResponse": {
				"type": "object",
				"properties": {
					"following": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Follow"
						}
					},
					"pagination": {
						"type": "object",
						"properties": {
							"page": {
								"type": "integer"
							},
							"limit": {
								"type": "integer"
							},
							"total": {
								"type": "integer"
							},
							"totalPages": {
								"type": "integer"
							}
						}
					}
				}
			},
...
			"Post": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"caption": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"userAvatar": {
						"type": "string"
					},
					"media": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/MediaDto"
						}
					},
					"location": {
						"$ref": "#/components/schemas/LocationDto"
					},
					"hashtags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"counts": {
						"$ref": "#/components/schemas/PostCounts"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"CoordinatesDto": {
				"type": "object",
				"properties": {
					"latitude": {
						"type": "number"
					},
					"longitude": {
						"type": "number"
					}
				}
			},
			"LocationDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"coordinates": {
						"$ref": "#/components/schemas/CoordinatesDto"
					}
				}
			},
			"MediaDto": {
				"type": "object",
				"properties": {
					"type": {
						"type": "string",
						"enum": ["image", "video"]
					},
					"url": {
						"type": "string"
					},
					"thumbnailUrl": {
						"type": "string"
					},
					"width": {
						"type": "number"
					},
					"height": {
						"type": "number"
					},
					"duration": {
						"type": "number"
					}
				}
			},
...
			"UpdatePostDto": {
				"type": "object",
				"properties": {
					"caption": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/LocationDto"
					},
					"hashtags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"isCommentsEnabled": {
						"type": "boolean"
					},
					"isLikesVisible": {
						"type": "boolean"
					}
				}
			},
			"PostCounts": {
				"type": "object",
				"properties": {
					"likes": {
						"type": "integer"
					},
					"comments": {
						"type": "integer"
					}
				}
			},
			"PostResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/Post"
					}
				}
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
...
	"security": [{ "bearerAuth": [] }]
}